requirements.txt
requests
flask
peewee
pillow
gunicorn
openai
python-dotenv